Removes unused addressBook container
